requirements.txt
requests==2.31.0
sqlalchemy==2.0.25
pandas==2.1.4
schedule==1.2.1
python-dotenv==1.0.1
streamlit==1.30.0
altair==5.2.0